+++ lib/command.js
@@ -24,9 +24,6 @@ class Command {
      */
     this.yargs = yargs;
 
-    this.version = pkg.version;
-    this.usage = '';
-
     /**
      * command name, just `start` or advance usage `get <source> [proxy]`
      * @type {String}
@@ -75,7 +72,8 @@ class Command {
 
   /**
    * load sub commands
-   * @param {...String} args - the command directory
+   * @param {...String} args - the command directory, args will be join by `path.join`
+   * @example `loadCommand(__dirname, 'command')`
    */
   loadCommand(...args) {
     const fullPath = path.join(...args);
@@ -95,7 +93,8 @@ class Command {
 
   /**
    * add sub command
-   * @param {...String} args - special file path (must contains ext)
+   * @param {...String} args - special file path (must contains ext), args will be join by `path.join`
+   * @example `loadCommand(__dirname, 'test_command.js')`
    */
   addCommand(...args) {
     const filePath = path.join(...args);
@@ -130,7 +129,7 @@ class Command {
     this.name = this.name || pkg.name;
     this.yargs
       .wrap(120)
-      .version(this.version)
+      .version()
       .help()
       .alias('h', 'help')
       .group([ 'help', 'version' ], 'Global Options:')
@@ -138,7 +137,7 @@ class Command {
       // .demandCommand(1)
       // .recommendCommands()
       // .epilogue(`for more information, find our manual at ${pkg.homepage}`)
-      .usage(this.usage)
+      .usage(this.usage || '')
       .options(this.options);
 
     // register to yargs
