+++ lib/helper.js
@@ -22,8 +22,10 @@ function gracefull(proc) {
         process.exit(0);
       });
     });
+
     process.once('exit', () => {
       // had test at my-helper.test.js, but coffee can't collect coverage info.
+      /* istanbul ignore next */
       for (const child of childs) {
         debug('kill child %s with %s', child.pid, signal);
         child.kill(signal);
