+++ test/fixtures/async-bin/command/sleep.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const Command = require('../../../../..').Command;
+const Command = require('..');
 
 class SleepCommand extends Command {
   constructor() {
