+++ test/fixtures/async-bin/scripts/echo_script.js
