+++ test/fixtures/my-bin/command/error.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const Command = require('../../../../..');
+const Command = require('..');
 
 class ErrorCommand extends Command {
   constructor() {
