+++ test/fixtures/my-bin/command/start-cluster
