+++ test/fixtures/my-bin/command/start.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const Command = require('../../../../..');
+const Command = require('..');
 const path = require('path');
 
 class StartCommand extends Command {
