+++ test/fixtures/my-git/index.js
@@ -1,3 +1,18 @@
 'use strict';
 
-module.exports = require('./lib/program');
+const BaseCommand = require('../../..');
+const pkg = require('./package.json');
+
+class Command extends BaseCommand {
+  start() {
+    this.name = pkg.name;
+    this.usage = `Usage: ${this.name} <command> [options]`;
+
+    // load sub command
+    this.loadCommand(__dirname, 'command');
+
+    super.start();
+  }
+}
+
+module.exports = Command;
