+++ test/fixtures/my-helper/command/echo.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const Command = require('../../../../..');
+const Command = require('..');
 
 class EchoCommand extends Command {
   constructor() {
