+++ test/fixtures/my-helper/command/install.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const path = require('path');
-const Command = require('../../../../..');
+const Command = require('..');
 
 class InstallCommand extends Command {
   constructor() {
