+++ test/fixtures/my-helper/index.js
@@ -1,3 +1,22 @@
 'use strict';
 
-module.exports = require('./lib/program');
+const BaseCommand = require('../../..');
+const pkg = require('./package.json');
+const helper = require('./lib/helper');
+
+class Command extends BaseCommand {
+  start() {
+    this.name = pkg.name;
+    this.usage = `Usage: ${this.name} <command> [options]`;
+
+    // load sub command
+    this.loadCommand(__dirname, 'command');
+
+    // custom helper
+    Object.assign(this.helper, helper);
+
+    super.start();
+  }
+}
+
+module.exports = Command;
