+++ test/fixtures/my-helper/scripts/error_script.js
