+++ test/fixtures/my-helper/scripts/loop_script.js
