+++ test/fixtures/my-helper/scripts/test_script.js
