+++ test/fixtures/single-bin/bin/single-bin.js
@@ -2,5 +2,5 @@
 
 'use strict';
 
-const Program = require('../lib/program');
-new Program().start();
+const Command = require('..');
+new Command().start();
