+++ test/fixtures/single-bin/index.js
@@ -1,3 +1,25 @@
 'use strict';
 
-module.exports = require('./lib/program');
+const Command = require('../../..');
+const pkg = require('./package.json');
+
+class Program extends Command {
+  start() {
+    this.name = pkg.name;
+
+    this.options = {
+      baseDir: {
+        description: 'target directory',
+      },
+    };
+
+    super.start();
+  }
+
+  run(context) {
+    console.log('run default command at %s', context.argv.baseDir);
+    this.showHelp();
+  }
+}
+
+module.exports = Program;
