--- test/fixtures/single-bin/lib/program.js
@@ -1,25 +1,0 @@
-'use strict';
-
-const Command = require('../../../..');
-const pkg = require('../package.json');
-
-class Program extends Command {
-  constructor() {
-    super();
-    this.name = pkg.name;
-    this.version = pkg.version;
-
-    this.options = {
-      baseDir: {
-        description: 'target directory',
-      },
-    };
-  }
-
-  run(context) {
-    console.log('run default command at %s', context.argv.baseDir);
-    this.showHelp();
-  }
-}
-
-module.exports = Program;
